Add a basic "single batch" mode
This was for running a small pilot test. If `SINGLE_BATCH_MODE` is set
to True, it bypasses all the normal batch assignment logic and simply
assigns every user to the given batch ID.

It currently doesn't check if a user has already completed the batch, if
this functionality needs to be used again it might be worth fixing that.