Change redirect when no batches available
Instead of redirecting to the "Thank you" page users should now be
redirected to a page without the completion code - there were cases
where people were attempting to perform the evaluations but would get
routed straight to the "Thank you" page becase there were no batches to
assign to them.

There's now a dedicated page for this case which explains the study is
complete.